refactor(ses): Move makeLRUCacheMap to a dedicated file
It should eventually be an endo export, possibly in its own package
(just not in `common`, which itself depends upon `ses`).
